Renamed load.py to client_weights.py
